Time: 83 ms (88.93%) | Memory: 79.1 MB (64.00%) - LeetSync

+++ 2616-maximal-score-after-applying-k-operations/maximal-score-after-applying-k-operations.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long maxKelements(vector<int>& nums, int k) {
-        priority_queue<int> pq(nums.begin(), nums.end());
+        priority_queue<double> pq(nums.begin(), nums.end());
         long long score = 0;
 
         while(k--){
@@ -10,6 +10,7 @@ public:
             score += temp;
             pq.push(ceil(temp/3));
         }
+
         return score;;
     }
 };
